updated to net9.0 and enabled nullable annotations.  remove Newtonsoft.Json dependency and use System.Text.Json instead.  further refactoring and moving files around.
