--- src/Harmony/RunTimePatch.cs
@@ -1,187 +1,0 @@
-﻿using System.Text.RegularExpressions;
-using Newtonsoft.Json;
-using HarmonyLib;
-
-//original work done by KK
-//modifications for patching weaponType and headshots added from Mustached_Maniac
-
-namespace _7DTDWebsockets.patchs
-{
-    public sealed class PlayerDmgEvent
-    {
-        public readonly Player player;
-        public readonly string cause;
-        public readonly int damage;
-
-        public PlayerDmgEvent(Player player, string cause, int damage)
-        {
-            this.player = player;
-            this.cause = cause;
-            this.damage = damage;
-        }
-    }
-
-    public sealed class PlayerKillEntityEvent
-    {
-        public readonly Player player;
-        public readonly string entity;
-        public readonly bool animal;
-        public readonly bool zombie;
-        public readonly string weaponType;
-        public readonly bool headshot;
-
-        public PlayerKillEntityEvent(Player player, string entity, bool animal, bool zombie, string weaponType, bool headshot)
-        {
-            this.player = player;
-            this.entity = entity;
-            this.animal = animal;
-            this.zombie = zombie;
-            this.weaponType = weaponType;
-            this.headshot = headshot;
-        }
-    }
-
-    public sealed class PlayerEntityEvent
-    {
-        public readonly Player player;
-        public readonly string entity;
-
-        public PlayerEntityEvent(Player player, string entity)
-        {
-            this.player = player;
-            this.entity = entity;
-        }
-    }
-
-    public sealed class PlayerOnlyEvent
-    {
-        public readonly Player player;
-
-        public PlayerOnlyEvent(Player player)
-        {
-            this.player = player;
-        }
-    }
-
-    public sealed class PlayerDeathEvent
-    {
-        public readonly Player player;
-
-        public PlayerDeathEvent(Player player)
-        {
-            this.player = player;
-        }
-    }
-
-    [HarmonyPatch(typeof(EntityAlive), "SetDead")]
-    public static /*partial*/ class PatchEntityDeath
-    {
-        // if the [GeneratedRegex] attribute can't be used, still prefer static, compiled Regex fields
-        private static readonly Regex animalEntityNameRegex = new Regex(@"(animal)", RegexOptions.Compiled);
-        private static readonly Regex zombieEntityNameRegex = new Regex(@"(zombie)", RegexOptions.Compiled);
-
-        // using the [GeneratedRegex] attribute, the regex is compiled at build time!
-        //[GeneratedRegex(@"(animal)")]
-        //private static partial Regex AnimalEntityNameRegex();
-
-        //[GeneratedRegex(@"(zombie)")]
-        //private static partial Regex ZombieEntityNameRegex();
-
-        public static bool IsHeadshot; // TODO: this seems very, very wrong
-
-        public static bool Prefix(EntityAlive __instance)
-        {
-            if (__instance is EntityPlayer)
-            {
-                return true;
-            }
-
-            var obj = Traverse.Create(__instance).Field("entityThatKilledMe").GetValue();
-            if (obj == null)
-            {
-                return true;
-            }
-
-            if (!(obj is EntityPlayer player))
-            {
-                return true;
-            }
-
-            string entityNameLower = __instance.GetDebugName().ToLower();
-
-            bool isAnimal = false;
-            bool isZombie = false;
-
-            if (entityNameLower.Contains("animal"))
-            {
-                isAnimal = true;
-                animalEntityNameRegex.Replace(entityNameLower, "");
-            }
-
-            if (entityNameLower.Contains("zombie"))
-            {
-                isZombie = true;
-                zombieEntityNameRegex.Replace(entityNameLower, "");
-            }
-
-            API.Send("PlayerKillEntity", JsonConvert.SerializeObject(new PlayerKillEntityEvent(new Player(player), entityNameLower, isAnimal, isZombie, player.inventory.holdingItem.Name, IsHeadshot)));
-
-            if (isZombie)
-            {
-                API.Send("PlayerKillZombie", JsonConvert.SerializeObject(new PlayerEntityEvent(new Player(player), entityNameLower)));
-            }
-            else if (isAnimal)
-            {
-                API.Send("PlayerKillAnimal", JsonConvert.SerializeObject(new PlayerEntityEvent(new Player(player), entityNameLower)));
-            }
-
-            return true;
-        }
-    }
-
-    [HarmonyPatch(typeof(EntityAlive), "damageEntityLocal")]
-    public static class PatchDamageEntityLocal
-    {
-        public static void Postfix(DamageResponse __result)
-        {
-            PatchEntityDeath.IsHeadshot = __result.HitBodyPart == EnumBodyPartHit.Head; // TODO: this seems very, very wrong
-        }
-    }
-
-    [HarmonyPatch]
-    public static class DamagePatches
-    {
-        [HarmonyPrefix]
-        [HarmonyPatch(typeof(NetPackageDamageEntity), "ProcessPackage")]
-        public static void DamageEntityPacketProccessPrefix(NetPackage __instance, World _world, GameManager _callbacks)
-        {
-            // TODO: seems like maybe this should be pattern matching on the type?
-            if (NetPackageManager.GetPackageName(__instance.PackageId) != nameof(NetPackageDamageEntity))
-            {
-                return;
-            }
-
-            NetPackageDamageEntity damage = (NetPackageDamageEntity)__instance;
-            Entity entity = _world.GetEntity(damage.entityId);
-            if (entity == null || !(entity is EntityPlayer entityPlayer))
-            {
-                return;
-            }
-
-            API.Send("PlayerDamage", JsonConvert.SerializeObject(new PlayerDmgEvent(new Player(entityPlayer), damage.damageTyp.ToString(), damage.strength)));
-        }
-
-        [HarmonyPrefix]
-        [HarmonyPatch(typeof(EntityAlive), "DamageEntity")]
-        public static void EntityAliveDamagePrefix(EntityAlive __instance, DamageSource _damageSource, int _strength, bool _criticalHit, float _impulseScale = 1f)
-        {
-            if (!(__instance is EntityPlayer player))
-            {
-                return;
-            }
-
-            API.Send("PlayerDamage", JsonConvert.SerializeObject(new PlayerDmgEvent(new Player(player), _damageSource.damageType.ToString(), _strength)));
-            return;
-        }
-    }
-}
