+++ src/Player.cs
@@ -1,8 +1,8 @@
 ﻿namespace _7DTDWebsockets
 {
-    public class Player
+    public sealed class Player
     {
-        public string name;
+        public readonly string name;
 
         public Player(ClientInfo clientInfo)
         {
