--- src/packages.config
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Lib.Harmony" version="2.3.3" targetFramework="net48" />
-  <package id="Newtonsoft.Json" version="13.0.1" targetFramework="net48" />
-  <package id="WebSocketSharp" version="1.0.3-rc11" targetFramework="net48" />
-</packages>
